Refactor and improv code readability
