+++ src/shaders/fragment.glsl
@@ -1,0 +1,6 @@
+varying highp vec4 vColor;
+varying highp vec3 vLighting;
+
+void main() {
+    gl_FragColor = vec4(vColor.rgb * vLighting, vColor.a);
+}
