+++ webpack.config.js
@@ -16,16 +16,20 @@ module.exports = {
         test: /\.(png|jpe?g|gif|mp4)$/i,
         use: [
           {
-            loader: 'url-loader',
+            loader: "url-loader",
             options: {
               limit: 8192,
-              name: '[name].[ext]',
-              outputPath: 'images',
-              publicPath: 'images/',
+              name: "[name].[ext]",
+              outputPath: "images",
+              publicPath: "images/",
               esModule: false,
             },
           },
         ],
+      },
+      {
+        test: /\.glsl/i,
+        use: ["raw-loader"],
       },
     ],
   },
